SearchFast V1
This version has flawed Highlighting and doesn't yet allow to go to the files. 
Ideas for Future Versions include:
Improve highlighting of words with actual color
add context lines and be allowed to choose the amount of context you want
opens file
opens a view of file where you can see a preview of the full file with the searchable text being highlighted. 
Have an ETA of when the search will be done
have a progress bar
have a counter which shows which files are being searched. 
have a counter at the beginning of the script which counts how many files it needs to go through so you can estimate a progress bar
have a equation which will tell you how much time you saved
have a little notice which will tell you how fast the results were.